backend/requirements.txt
fastapi
uvicorn
pydantic
PyYAML
lunar-python
